Implement dungeon view window
Moved drawing of dungeons into a separate class to allow multiple
dungeons to be handled simultaneously. I also consolidated the drawing
of chars into the dungeon and dungeon_view. Objects, tiles and
characters should be able to provide their char and color info but don't
need to know about the console.